+++ src/test/java/com/xiaoleilu/hutool/test/Test.java
@@ -7,6 +7,6 @@ package com.xiaoleilu.hutool.test;
  *
  */
 public class Test {
-	public static void main(String[] args) throws InterruptedException {
+	public static void main(String[] args) throws Exception {
 	}
 }